Ensure editors are updated when backlinks are updated or deleted
