+++ editor/backlinks/deleteBacklinks.ts
@@ -3,8 +3,7 @@ import { ElementType } from 'types/slate';
 import { Note } from 'types/supabase';
 import supabase from 'lib/supabase';
 import { store } from 'lib/store';
-import activeEditorsStore from 'lib/activeEditorsStore';
-import createNotabaseEditor from 'editor/createEditor';
+import { getActiveOrTempEditor } from 'lib/activeEditorsStore';
 import { computeLinkedBacklinks } from './useBacklinks';
 
 /**
@@ -22,12 +21,7 @@ const deleteBacklinks = async (noteId: string) => {
       continue;
     }
 
-    // Get editor from active editors if it exists, or create a new one
-    let editor = activeEditorsStore.getActiveEditor(backlink.id);
-    if (!editor) {
-      editor = createNotabaseEditor();
-      editor.children = note.content;
-    }
+    const editor = getActiveOrTempEditor(backlink.id, note.content);
 
     Transforms.unwrapNodes(editor, {
       at: [],
